+++ src/BugsnagUnity/Client.cs
@@ -397,5 +397,11 @@ namespace BugsnagUnity
       // propagate the change to the native property also
       NativeClient.SetAutoNotify(autoNotify);
     }
+
+    public void SetAutoDetectAnrs(bool autoDetectAnrs)
+    {
+      // Set the native property
+      NativeClient.SetAutoDetectAnrs(autoDetectAnrs);
+    }
   }
 }